Fixes to the Set Location View in the Event Creation Process.  Consolidated buttons and fixed borderwidth/color issues. Updated image assets.  Cleaned up code comments and removed part of old code. Automatic sign-in implemented along with comments.  Event Detail loading is faster and more similar to Facebook where the user sees a preview of what's coming and details fade in as they are loaded.

+++ EVNTR/FilterEventsVC.h
@@ -12,7 +12,7 @@
 
 @interface FilterEventsVC : UIViewController
 
-@property (nonatomic) int selectedFilterDistance;
+@property (nonatomic) float selectedFilterDistance;
 @property (nonatomic, strong) id <EVNFilterProtocol> delegate;
 
 
@@ -20,6 +20,6 @@
 
 @protocol EVNFilterProtocol <NSObject>
 
-- (void) completedFiltering:(int)radius;
+- (void) completedFiltering:(float)radius;
 
 @end
